Issue #394 -- Refactor the URI canonicalization into the Encoder class.
