+++ src/main/java/org/owasp/esapi/Validator.java
@@ -709,17 +709,6 @@ public interface Validator {
 	boolean isValidURI(String context, String input, boolean allowNull);
 
 	/**
-	 * 
-	 * Get a version of the input URI that will be safe to run regex and other validations against.  
-	 * It is not recommended to persist this value as it will transform user input.  This method 
-	 * will not test to see if the URI is RFC-3986 compliant.
-	 * 
-	 * @param input
-	 * @return
-	 */
-	public String getCanonicalizedURI(URI dirtyUri);
-	
-	/**
 	 * Will return a {@code URI} object that will represent a fully parsed and legal URI
 	 * as specified in RFC-3986.  
 	 *
